fix: bug ao criar um usuario apos algum id deletado, erro de duplicação de key corrigido

+++ src/main/java/br/com/douglas/travelbooking/controllers/ClinicaController.java
@@ -22,7 +22,7 @@ public class ClinicaController {
 	@PostMapping
 	public ResponseEntity<Clinica> insert(@RequestBody Clinica objClinicaDto) {
 	 
-	    int id = (int) service.getClinicaCount();
+	    int id = service.getClinicaCount();
 	    String nome = objClinicaDto.getNomeClinica();
 	    Clinica response = new Clinica(id, nome);
 	    Clinica objUser = service.fromDTO(objClinicaDto);
